[FIX] edi: split gateway test cases
