+++ addons/edi/tests/__init__.py
@@ -2,8 +2,10 @@
 
 from . import common
 from . import test_gateway
+from . import test_edi_connection_local
 from . import test_edi_connection_mail
 from . import test_edi_connection_sftp
+from . import test_edi_connection_xmlrpc
 from . import test_edi_document
 from . import test_edi_issue
 from . import test_edi_transfer
