+++ addons/edi/tests/test_gateway.py
@@ -1,5 +1,4 @@
 from . import common
-import os.path
 
 from odoo import fields
 
@@ -108,114 +107,3 @@ class TestEDIGateway(common.BaseEDI):
         self.assertEqual(new_transfer, transfer)
         # transfer has one attachment
         self.assertEqual(len(transfer.output_ids), 1)
-
-
-from base64 import b64encode
-
-class TestEDIXMLRPCGateway(TestEDIGateway):
-
-    @classmethod
-    def setUpClass(cls):
-        super(TestEDIXMLRPCGateway, cls).setUpClass()
-
-    def setUp(self):
-        super(TestEDIXMLRPCGateway, self).setUp()
-
-    @classmethod
-    def _create_xmlrpc_gateway(cls):
-        model = cls.env.ref('edi.model_edi_connection_xmlrpc')
-        return cls._create_gateway(name='Test XMLRPC Gateway',
-                                   model_id=model.id)
-
-    @classmethod
-    def _create_xmlrpc_path(cls, gateway):
-        data = {
-            'name': 'Local path for %s' % gateway.name,
-            'gateway_id': gateway.id,
-            'glob': '*.txt',
-            'path': 'files',
-            'age_window': 24,
-        }
-        return cls._create_path(**data)
-
-    @classmethod
-    def _create_xmlrpc_conn(cls):
-        prf_filename = os.path.join(os.path.dirname(os.path.abspath(__file__)),
-                                    'files', 'hello_world.txt')
-
-        conn = {'files': [{
-            'name': 'hello_world.txt',
-            'data': b64encode(open(prf_filename, 'rb').read()),
-        }]}
-
-        return conn
-
-    def test02_gateway_do_transfer_no_path(self):
-
-        gateway = self._create_xmlrpc_gateway()
-        conn = self._create_xmlrpc_conn()
-        self._test_do_transfer_no_path(gateway, conn=conn)
-
-
-    def test03_gateway_do_transfer_receive(self):
-
-        gateway = self._create_xmlrpc_gateway()
-        path = self._create_xmlrpc_path(gateway)
-
-        conn = self._create_xmlrpc_conn()
-
-        self._test_do_transfer_receive(gateway, conn=conn)
-
-
-class TestEDILocalGateway(TestEDIGateway):
-
-    @classmethod
-    def setUpClass(cls):
-        super(TestEDILocalGateway, cls).setUpClass()
-
-    def setUp(self):
-        super(TestEDILocalGateway, self).setUp()
-
-    def test01_gateway_action_test(self):
-
-        gateway = self._create_local_gateway()
-        self._create_local_path(gateway=gateway,
-                                allow_send=False)
-        self._test_action_test(gateway)
-
-    def test02_gateway_do_transfer_no_path(self):
-
-        gateway = self._create_local_gateway()
-        self._test_do_transfer_no_path(gateway)
-
-    def test03_gateway_do_transfer_receive(self):
-
-        gateway = self._create_local_gateway()
-        self._create_local_path(gateway=gateway,
-                                allow_send=False)
-        self._test_do_transfer_receive(gateway)
-
-    def test04_gateway_do_transfer_receive_old_file_ignored(self):
-
-        gateway = self._create_local_gateway()
-        self._create_local_path(gateway=gateway,
-                                touch_files=False,
-                                folder='files/old',
-                                age_window=0.01)
-        self._test_do_transfer_receive_old_file_ignored(gateway)
-
-    def test05_gateway_do_transfer_send(self):
-
-        gateway = self._create_local_gateway()
-        self._create_local_path(gateway=gateway,
-                                folder='files/out',
-                                glob='*.out',
-                                allow_receive=False,
-                                doc_type_ids=[(6, 0, [self.document_type_unknown.id])],
-                                )
-        self._test_do_transfer_send(gateway)
-
-    def test06_gateway_action_transfer_no_path(self):
-
-        gateway = self._create_local_gateway()
-        self._test_action_transfer_no_path(gateway)
